Many such cases

+++ Blok34.Web/Controllers/EventsController.cs
@@ -1,4 +1,5 @@
 ﻿using Blok34.Domain.DomainModels;
+using Blok34.Domain.Enums;
 using Blok34.Service.Implementation;
 using Blok34.Service.Interface;
 using Blok34.Web.Data;
@@ -9,6 +10,7 @@ using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace Blok34.Web.Controllers
@@ -19,14 +21,14 @@ namespace Blok34.Web.Controllers
         private readonly IEventService _eventService;
         private readonly IVenueService _venueService;
         private readonly IEventAttendanceService _eventAttendanceService;
+        private readonly IUserService _userService;
 
-        public EventsController(IEventService eventService,
-            IVenueService venueService,
-            IEventAttendanceService eventAttendanceService)
+        public EventsController(IEventService eventService, IVenueService venueService, IEventAttendanceService eventAttendanceService, IUserService userService)
         {
             _eventService = eventService;
             _venueService = venueService;
             _eventAttendanceService = eventAttendanceService;
+            _userService = userService;
         }
 
         // GET: Events
@@ -46,6 +48,36 @@ namespace Blok34.Web.Controllers
 
             return View(@event);
         }
+
+        [Authorize]
+        [HttpPost]
+        public IActionResult UpdateAttendance(Guid eventId, AttendanceStatus status)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            if (status == AttendanceStatus.Attending)
+            {
+                _eventAttendanceService.AttendEvent(eventId, userId);
+            }
+            else if (status == AttendanceStatus.Interested)
+            {
+                _eventAttendanceService.MarkInterested(eventId, userId);
+            }
+
+            return Ok();
+        }
+
+        [Authorize]
+        [HttpPost]
+        public IActionResult CancelAttendance(Guid eventId)
+        {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            _eventAttendanceService.RemoveAttendance(eventId, userId);
+
+            return Ok();
+        }
+
 
         // GET: Events/Create
         public IActionResult Create()
@@ -68,29 +100,59 @@ namespace Blok34.Web.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Create([Bind("Title,Description,StartDate,EndDate,VenueId,Category,CreatedByUserId,Id")] Event @event)
         {
+            ModelState.Remove("Venue");
+
+            var venue = _venueService.GetVenueById(@event.VenueId);
+            if (venue != null)
+            {
+                @event.Venue = venue;
+            }
+
             if (ModelState.IsValid)
             {
+                
                 _eventService.Insert(@event);
                 return RedirectToAction(nameof(Index));
             }
+
+            var userId = User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+            var venues = _venueService.GetAllVenues()
+                            .Where(v => v.IsPublic || v.VenueManagerId == userId)
+                            .ToList();
+
+            // Provide the list again and set selected value to restore hidden select
+            ViewData["VenueId"] = new SelectList(venues, "Id", "Name", @event?.VenueId);
+
+
             return View(@event);
         }
 
         // GET: Events/Edit/5
         public IActionResult Edit(Guid id)
         {
-            var userId = User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            var @event = _eventService.GetEventById(id);
+
+            if (@event.CreatedByUserId != userId)
+            {
+                return Forbid();
+            }
 
             var venues = _venueService.GetAllVenues()
             .Where(v => v.IsPublic || v.VenueManagerId == userId)
             .ToList();
 
-            
-            var @event = _eventService.GetEventById(id);
             if (@event == null)
             {
                 return NotFound();
             }
+
+            if (DateTime.UtcNow >= @event.StartDate)
+            {
+                return Forbid();
+            }
+
 
             ViewData["VenueId"] = new SelectList(venues, "Id", "Name", @event.VenueId);
             return View(@event);
@@ -108,19 +170,44 @@ namespace Blok34.Web.Controllers
                 return NotFound();
             }
 
-            _eventService.Update(@event);
-            return RedirectToAction(nameof(Index));
-          //  return View(@event);
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (@event.CreatedByUserId != userId)
+            {
+                return Forbid();
+            }
+
+            if (ModelState.IsValid)
+            {
+                _eventService.Update(@event);
+                return RedirectToAction(nameof(Index));
+            }
+
+            var venues = _venueService.GetAllVenues()
+                .Where(v => v.IsPublic || v.VenueManagerId == userId)
+                .ToList();
+
+            ViewData["VenueId"] = new SelectList(venues, "Id", "Name", @event.VenueId);
+            return View(@event);
         }
 
         // GET: Events/Delete/5
         public IActionResult Delete(Guid id)
         {
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+
             var @event = _eventService.GetEventById(id);
             if (@event == null)
             {
                 return NotFound();
             }
+
+            if (@event.CreatedByUserId != userId)
+            {
+                return Forbid(); // or Unauthorized()
+            }
+
 
             return View(@event);
         }
@@ -130,6 +217,15 @@ namespace Blok34.Web.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult DeleteConfirmed(Guid id)
         {
+
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var @event = _eventService.GetEventById(id);
+
+            if (@event.CreatedByUserId != userId)
+            {
+                return Forbid();
+            }
+
             _eventService.DeleteById(id);
 
             return RedirectToAction(nameof(Index));
